ref: move storage to its own crate
