+++ src/processor/snapshot/exporter.rs
@@ -2,19 +2,20 @@ use std::path::{Path, PathBuf};
 
 use ethers::types::{U256, U64};
 use eyre::Result;
+use state_reconstruct_storage::{
+    types::{
+        Proto, SnapshotFactoryDependencies, SnapshotFactoryDependency, SnapshotHeader,
+        SnapshotStorageLogsChunk, SnapshotStorageLogsChunkMetadata,
+    },
+    InnerDB, FACTORY_DEPS, INDEX_TO_KEY_MAP,
+};
 
-use super::{
-    database::{self, SnapshotDB},
-    types::{SnapshotFactoryDependency, SnapshotHeader},
-    DEFAULT_DB_PATH, SNAPSHOT_FACTORY_DEPS_FILE_NAME_SUFFIX, SNAPSHOT_HEADER_FILE_NAME,
-};
-use crate::processor::snapshot::types::{
-    Proto, SnapshotFactoryDependencies, SnapshotStorageLogsChunk, SnapshotStorageLogsChunkMetadata,
-};
+use super::{DEFAULT_DB_PATH, SNAPSHOT_HEADER_FILE_NAME};
+use crate::processor::snapshot::SNAPSHOT_FACTORY_DEPS_FILE_NAME_SUFFIX;
 
 pub struct SnapshotExporter {
     basedir: PathBuf,
-    database: SnapshotDB,
+    database: InnerDB,
 }
 
 impl SnapshotExporter {
@@ -24,7 +25,7 @@ impl SnapshotExporter {
             None => PathBuf::from(DEFAULT_DB_PATH),
         };
 
-        let database = SnapshotDB::new_read_only(db_path)?;
+        let database = InnerDB::new_read_only(db_path)?;
         Ok(Self {
             basedir: basedir.to_path_buf(),
             database,
@@ -61,7 +62,7 @@ impl SnapshotExporter {
     fn export_factory_deps(&self, header: &mut SnapshotHeader) -> Result<()> {
         tracing::info!("Exporting factory dependencies...");
 
-        let storage_logs = self.database.cf_handle(database::FACTORY_DEPS).unwrap();
+        let storage_logs = self.database.cf_handle(FACTORY_DEPS).unwrap();
         let mut iterator = self
             .database
             .iterator_cf(storage_logs, rocksdb::IteratorMode::Start);
@@ -93,7 +94,7 @@ impl SnapshotExporter {
         let num_logs = self.database.get_last_repeated_key_index()?;
         tracing::info!("Found {num_logs} logs.");
 
-        let index_to_key_map = self.database.cf_handle(database::INDEX_TO_KEY_MAP).unwrap();
+        let index_to_key_map = self.database.cf_handle(INDEX_TO_KEY_MAP).unwrap();
         let mut iterator = self
             .database
             .iterator_cf(index_to_key_map, rocksdb::IteratorMode::Start);
