+++ state-reconstruct-fetcher/src/types/common.rs
@@ -1,7 +1,8 @@
 //! A collection of functions that get reused throughout format versions.
 use ethers::{abi, types::U256};
+use state_reconstruct_storage::PackingType;
 
-use super::{L2ToL1Pubdata, PackingType, ParseError};
+use super::{L2ToL1Pubdata, ParseError};
 use crate::constants::zksync::{
     L2_TO_L1_LOG_SERIALIZE_SIZE, LENGTH_BITS_OFFSET, OPERATION_BITMASK,
 };
