+++ state-reconstruct-storage/proto/snapshot.proto
@@ -1,0 +1,23 @@
+syntax = "proto3";
+
+package protobuf;
+
+message SnapshotStorageLogsChunk {
+    repeated SnapshotStorageLog storage_logs = 1;
+}
+
+message SnapshotStorageLog {
+    optional bytes account_address = 1; // required; H160
+    optional bytes storage_key = 2; // required; H256
+    optional bytes storage_value = 3; // required; H256
+    optional uint32 l1_batch_number_of_initial_write = 4; // required
+    optional uint64 enumeration_index = 5; // required
+}
+
+message SnapshotFactoryDependencies {
+    repeated SnapshotFactoryDependency factory_deps = 1;
+}
+
+message SnapshotFactoryDependency {
+    optional bytes bytecode = 1; // required
+}
